평점보기

+++ users/views.py
@@ -2,7 +2,7 @@ from rest_framework.views import APIView
 from rest_framework.generics import get_object_or_404
 from rest_framework import status, permissions
 from rest_framework.response import Response
-from users.serializers import PetOwnerReviewCreateSerializer, PetSitterReviewCreateSerializer,PetOwnerReviewSerializer,PetSitterReviewSerializer
+from users.serializers import PetOwnerReviewCreateSerializer, PetSitterReviewCreateSerializer,PetOwnerReviewSerializer,PetSitterReviewSerializer,StarRatingSerializer
 from users.models import PetOwnerReview, PetSitterReview, User
 
 
@@ -16,6 +16,8 @@ class PetOwnerReviewView(APIView):
 
     # 후기 작성하기
     def post(self, request, user_id):
+        permission_classes = [permissions.IsAuthenticated]
+
         serializer = PetOwnerReviewCreateSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save(writer = request.user, owner_id = user_id)
@@ -35,17 +37,23 @@ class PetOwnerReviewDetailView(APIView):
     def put(self, request, user_id, review_id):
         ownerreview = get_object_or_404(PetOwnerReview, pk=review_id)
         serializer = PetOwnerReviewCreateSerializer(ownerreview, data=request.data)
-        if serializer.is_valid():
-            serializer.save(writer = request.user, owner_id = user_id)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.user == ownerreview.writer:
+            if serializer.is_valid():
+                serializer.save(writer = request.user, owner_id = user_id)
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'mesage': '권한이 없습니다!'}, status=status.HTTP_403_FORBIDDEN)
     
     # 후기 삭제하기
     def delete(self, request, user_id, review_id):
         ownerreview = get_object_or_404(PetOwnerReview, pk=review_id)
-        ownerreview.delete()
-        return Response({'mesage': '후기가 삭제되었습니다.'},status=status.HTTP_204_NO_CONTENT)
+        if request.user == ownerreview.writer:
+            ownerreview.delete()
+            return Response({'mesage': '후기가 삭제되었습니다.'},status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response({'mesage': '권한이 없습니다!'}, status=status.HTTP_403_FORBIDDEN)
     
 
 
@@ -59,6 +67,8 @@ class PetSitterReviewView(APIView):
 
     # 후기 작성하기
     def post(self, request, user_id):
+        permission_classes = [permissions.IsAuthenticated]
+
         serializer = PetSitterReviewCreateSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save(writer = request.user, sitter_id = user_id)
@@ -79,15 +89,28 @@ class PetSitterReviewDetailView(APIView):
     def put(self, request, user_id, review_id):
         sitterreview = get_object_or_404(PetSitterReview, pk=review_id)
         serializer = PetSitterReviewCreateSerializer(sitterreview, data=request.data)
-        if serializer.is_valid():
-            serializer.save(writer = request.user, sitter_id = user_id)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.user == sitterreview.writer:
+            if serializer.is_valid():
+                serializer.save(writer = request.user, sitter_id = user_id)
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'mesage': '권한이 없습니다!'}, status=status.HTTP_403_FORBIDDEN)
     
     # 후기 삭제하기
     def delete(self, request, user_id, review_id):
         sitterreview = get_object_or_404(PetSitterReview, pk=review_id)
-        sitterreview.delete()
-        return Response({'mesage': '후기가 삭제되었습니다.'},status=status.HTTP_204_NO_CONTENT)
+        if request.user == sitterreview.writer:
+            sitterreview.delete()
+            return Response({'mesage': '후기가 삭제되었습니다.'},status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response({'mesage': '권한이 없습니다!'}, status=status.HTTP_403_FORBIDDEN)
     
+
+# 유저의 후기 평점
+class StarRatingView(APIView):
+    def get(self, request, user_id):
+        user = get_object_or_404(User, pk = user_id)
+        serializer = StarRatingSerializer(user)
+        return Response(serializer.data, status=status.HTTP_200_OK)
